requirements.txt
discord.py>=2.4.0
aiohttp>=3.9.5
pandas>=2.2.2
python-dateutil>=2.9.0